Added json plugin using JSONPath
New action, see https://www.npmjs.com/package/JSONPath for details
- json:
    key1 : expression1
    key2 : expression2

By default it is evaluated against the response body (which is really
only useful if it was auto-converted from application/json), but you
can also specify a target property and it will pull that object from
the session as the evaluation target. Combining that with the -set action
can build some powerful object manipulation.