Deprecated custom_function actions to be replaced by "function" actions
Backwards compatible if you previously used a direct inline function,
but if your action used the expanded syntax with a .custom_function,
property on your action then you need to rename the property to
.function.

Synchronous functions are now supported. Previously you had to take
callback as the second parameter and call it when finished. Now, if your
function takes a single parameter, it will be considered synchronous.

Because of the optionality of the callback, I also changed the syntax
for the undocumented feature to import modules. Most people won't be
affected because the syntax was not yet documented.

To import modules in your custom function, add an imports property to
your action. This property should be a map of import names to module
paths. Each of these values will get evaluted in a require() statement
and then replaced back into the property map. The action itself will be
the "this" context when your function is called, so any of the required
modules can be referenced with this.imports.module_name. For example:

- imports:
    crypto: "crypto"
    my_module: "/path/to/my_module"
  function: !!js/function >
    function(context) {
      this.imports.my_module.myfunc();
    }