Simplified capture of HTTP headers
!!Breaking Change!!
Previously, with string-values in a headers action, if the string
did not match the pattern to be a regex (like "/pattern/gim"), then the
string was treated as an exact text match. Now, it will be considered as
the key for where to store the value of the header. This makes it much
easier to capture header values, which previously required an open
wild-card regex with a named group. The downside is that this is a
breaking change. The other downside is that the simple string/regex
shortcuts that can be used for top-level action shortcuts for text/regex
actions are not symmetrical inside of the headers element. Originally
I had thought that any top-level action might be able to be applied
to any header and so I was looking for symmetry with the construction
but the result was that simple things were harder than they needed to
be, in exchange for some unnecessary symmetry.

Old:
# To verify that a header value matches exact text
- headers:
   location: "this text would be matched"
# To capture the value of a header
-headers:
   location: /(:<capture_key>.*)/

New:
# To verify that a header value matches exact text
- headers:
   location: { text : "this text would be matched" }
# To capture the value of a header
- headers:
    location: capture_key