Fix for --init and --start flags w/invalid script flow
Because of the dynamically generated dummy “call” wrapper added to allow firstStep targets to be “called” instead of just invoked,
the processor was not able to find the current step in the index (because it had been dynamically generated) and so the indexOf returned -1.
When calculating the nextIndex -1 + 1 == 0, so it would start at the top of the script instead of continuing where it left off. Now we use paths to
lookup the current step and we copy the path from the target step to the dummy step.